Resolve correct CEL types consistently for all field types
